bugs fixed for btns of crud_employee_window.p/ it is worked properly with db now

+++ src/crud_employee_window.py
@@ -1,9 +1,11 @@
+from traceback import print_tb
 from PyQt6.QtWidgets import QDialog, QPushButton, QHBoxLayout, \
     QFrame, QVBoxLayout, QLineEdit, QMessageBox
 from PyQt6.QtCore import Qt
 from PyQt6.QtGui import QFont
 
-from services.users_employees_operations import create_employee, delete_employee
+from services.users_employees_operations import create_employee, \
+    delete_employee, get_employees
 
 
 class UserCrud(QDialog):
@@ -131,26 +133,51 @@ class UserCrud(QDialog):
         self.last_name_input_value = self.last_name_input.text()
         self.badge_input_value = self.badge_input.text()
 
-        create_employee(
-            self.name_input_value,
-            self.last_name_input_value,
-            self.badge_input_value
-        )
+        if (self.name_input_value and
+                self.last_name_input_value and self.badge_input_value):
+            create_employee(
+                self.name_input_value,
+                self.last_name_input_value,
+                self.badge_input_value
+            )
 
-        QMessageBox.information(
-            self,
-            "ثبت موفق",
-            "کارمند با موفقیت ثبت شد."
-        )
+            QMessageBox.information(
+                self,
+                "ثبت موفق",
+                "کارمند با موفقیت ثبت شد."
+            )
+        else:
+            QMessageBox.information(
+                self,
+                "ثبت ناموفق",
+                "مقدار خالی نمیتوان ذخیره کرد"
+            )
 
     def delete_employee(self):
-        delete_employee(
-            self.name_input_value,
-            self.last_name_input_value,
-            self.badge_input_value
-        )
+        self.name_input_value = self.name_input.text()
+        self.last_name_input_value = self.last_name_input.text()
+        self.badge_input_value = self.badge_input.text()
+
+        for emp, _ in get_employees():
+            if (emp.first_name == self.name_input_value and
+                    emp.last_name == self.last_name_input_value and 
+                    emp.badge == self.badge_input_value
+                ):
+                delete_employee(
+                    self.name_input_value,
+                    self.last_name_input_value,
+                    self.badge_input_value
+                )
+                QMessageBox.information(
+                    self,
+                    "حذف موفق",
+                    "کارمند با موفقیت حذف شد."
+                )
+                # i did not use else here casue i wanted to show just
+                # once if user not found with else it would be shown 
+                # several times(as much as records in db)
         QMessageBox.information(
-            self,
-            "حذف موفق",
-            "کارمند با موفقیت حذف شد."
-        )
+                    self,
+                    "حذف نا موفق",
+                    "کاربر مورد نظر پیدا نشد"
+                )
